Fetched Data of the best League players and added some styling to list of best players + some small css changes
removed LoggedinGuard for HomeComponent. fetched the challenger player data. sorted it by leaguepoints and displayed the 20 best in champion.html with the ngfor and slice. Also some styling was added for the list of the 20 best players and the text color was changed on all components to "blanchedalmond" to match the overall styling.